requirements.txt
click==7.1.2
requests==2.24.0
pyyaml==5.3.1
